refactor: 테스트용 목 클래스 레이어 변경
