+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -2,6 +2,7 @@ package kitchenpos.application;
 
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuGroupRepository;
+import kitchenpos.mock.InMemoryMenuGroupRepository;
 import kitchenpos.utils.fixture.MenuGroupFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
