+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -4,6 +4,9 @@ import kitchenpos.domain.MenuRepository;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.ProductRepository;
 import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.mock.FakePurgomalumClient;
+import kitchenpos.mock.InMemoryMenuRepository;
+import kitchenpos.mock.InMemoryProductRepository;
 import kitchenpos.utils.fixture.ProductFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
