+++ src/test/java/kitchenpos/mock/FakeKitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.mock;
 
 import kitchenpos.infra.KitchenridersClient;
 import org.springframework.stereotype.Component;
