+++ src/test/java/kitchenpos/mock/InMemoryProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.mock;
 
 import kitchenpos.domain.Product;
 import kitchenpos.domain.ProductRepository;
